+++ extensions/open-path/src/types/preference.ts
@@ -1,17 +1,15 @@
-import { Application, getPreferenceValues } from "@raycast/api";
+import { getPreferenceValues } from "@raycast/api";
 
-export interface Preference {
-  trimText: boolean;
-  isShowHud: boolean;
-  preferredTerminal: Application;
-  fileAction: string;
-  urlAction: string;
-  priorityDetection: string;
-  searchEngine: string;
-}
-
-export const { trimText, isShowHud, preferredTerminal, fileAction, urlAction, priorityDetection, searchEngine } =
-  getPreferenceValues<Preference>();
+export const {
+  trimText,
+  isShowHud,
+  preferredTerminal,
+  fileAction,
+  folderAction,
+  urlAction,
+  priorityDetection,
+  searchEngine,
+} = getPreferenceValues<Preferences>();
 
 export interface OpenInArguments {
   openIn: string;